Revert "Rework"
