--- data/public_servers.json
@@ -1,1 +1,0 @@
-{"1084731374344359966": null, "groups": []}
