+++ data/raid_pokemon
@@ -1,0 +1,1 @@
+[{'hp': 100, 'atk': 180, 'def': 160, 'spatk': 150, 'spdef': 90, 'speed': 90, 'type': ['Ground', 'Fire'], 'name': 'Primal Groudon', 'move': ['Ground', [100, 100], 'p', 'Earthquake']}, {'hp': 100, 'atk': 180, 'def': 160, 'spatk': 150, 'spdef': 90, 'speed': 90, 'type': ['Ground', 'Fire'], 'name': 'Primal Groudon', 'move': ['Ground', [120, 85], 'p', 'Precipice Blades']}, {'hp': 100, 'atk': 180, 'def': 160, 'spatk': 150, 'spdef': 90, 'speed': 90, 'type': ['Ground', 'Fire'], 'name': 'Primal Groudon', 'move': ['Ground', [90, 100], 's', 'Earth power']}, {'hp': 80, 'atk': 100, 'def': 123, 'spatk': 122, 'spdef': 120, 'speed': 80, 'type': ['Grass', 'Poison'], 'name': 'Mega venusaur', 'move': ['Grass', [120, 100], 's', 'Petal Dance']}, {'hp': 120, 'atk': 145, 'def': 120, 'spatk': 80, 'spdef': 90, 'speed': 105, 'type': ['Grass', ''], 'name': 'gigantamax Rillaboom', 'move': ['Grass', [150, 100], 'p', 'G-max Drum Solo']}, {'hp': 70, 'atk': 110, 'def': 75, 'spatk': 145, 'spdef': 85, 'speed': 145, 'type': ['Grass', 'Dragon'], 'name': 'Mega Sceptile', 'move': ['Grass', [120, 100], 's', 'Solar Beam']}, {'hp': 100, 'atk': 150, 'def': 90, 'spatk': 180, 'spdef': 160, 'speed': 90, 'type': ['Water', ''], 'name': 'Primal kyogre', 'move': ['Water', [150, 100], 's', 'Water spout']}, {'hp': 80, 'atk': 74, 'def': 74, 'spatk': 126, 'spdef': 116, 'speed': 60, 'type': ['Water', 'Fairy'], 'name': 'Primarina', 'move': ['Water', [195, 100], 's', 'oceanic operetta']}, {'hp': 100, 'atk': 180, 'def': 160, 'spatk': 150, 'spdef': 90, 'speed': 90, 'type': ['Ground', 'Fire'], 'name': 'Primal Groudon', 'move': ['Fire', [150, 100], 's', 'Eruption']}, {'hp': 78, 'atk': 104, 'def': 78, 'spatk': 159, 'spdef': 115, 'speed': 100, 'type': ['Fire', 'Flying'], 'name': 'Mega Charizard Y', 'move': ['Fire', [130, 90], 's', 'Overheat']}, {'hp': 137, 'atk': 113, 'def': 89, 'spatk': 137, 'spdef': 107, 'speed': 97, 'type': ['Psychic', 'Ghost'], 'name': 'Full Moon Lunala', 'move': ['Ghost', [200, 100], 's', 'Menacing Moonraze Maelstrom']}, {'hp': 106, 'atk': 190, 'def': 100, 'spatk': 154, 'spdef': 100, 'speed': 130, 'type': ['Psychic', 'Fighting'], 'name': 'Mega Mewtwo X', 'move': ['Fighting', [150, 100], 'p', 'Focus punch']}, {'hp': 80, 'atk': 185, 'def': 115, 'spatk': 40, 'spdef': 105, 'speed': 75, 'type': ['Bug', 'Fighting'], 'name': 'Mega heracross', 'move': ['Fighting', [150, 100], 'p', 'Focus punch']}, {'hp': 97, 'atk': 157, 'def': 127, 'spatk': 113, 'spdef': 109, 'speed': 77, 'type': ['Psychic', 'Steel'], 'name': 'Dusk necrozma', 'move': ['Steel', [200, 100], 'p', 'Searing Sunraze Smash']}, {'hp': 137, 'atk': 137, 'def': 107, 'spatk': 113, 'spdef': 89, 'speed': 97, 'type': ['Psychic', 'Steel'], 'name': 'Solgaleo', 'move': ['Steel', [200, 100], 'p', 'Searing Sunraze Smash']}, {'hp': 155, 'atk': 163, 'def': 163, 'spatk': 100, 'spdef': 85, 'speed': 54, 'type': ['Steel', ''], 'name': 'gigantamax Melmetal', 'move': ['Steel', [150, 100], 'p', 'G-max Meltdown']}, {'hp': 95, 'atk': 115, 'def': 90, 'spatk': 80, 'spdef': 90, 'speed': 60, 'type': ['Fire', 'Dark'], 'name': 'Incineroar', 'move': ['Dark', [180, 100], 'p', 'Malicious Moonsault']}, {'hp': 120, 'atk': 150, 'def': 120, 'spatk': 83, 'spdef': 80, 'speed': 117, 'type': ['Fighting', 'Dark'], 'name': 'gigantamax single strike Urshifu', 'move': ['Dark', [150, 100], 'p', 'G-max One Blow']}, {'hp': 115, 'atk': 140, 'def': 85, 'spatk': 115, 'spdef': 95, 'speed': 80, 'type': ['Dark', 'Fairy'], 'name': 'gigantamax Grimmsnarl', 'move': ['Dark', [150, 100], 'p', 'G-max Snooze']}, {'hp': 100, 'atk': 165, 'def': 150, 'spatk': 85, 'spdef': 130, 'speed': 50, 'type': ['Psychic', 'Ice'], 'name': 'Ice Rider Calyrex', 'move': ['Ice', [130, 100], 'p', 'Glacial Lance']}, {'hp': 125, 'atk': 170, 'def': 100, 'spatk': 120, 'spdef': 90, 'speed': 95, 'type': ['Dragon', 'Ice'], 'name': 'Black Kyurem', 'move': ['Ice', [140, 90], 'p', 'Freeze Shock']}, {'hp': 125, 'atk': 120, 'def': 90, 'spatk': 170, 'spdef': 100, 'speed': 95, 'type': ['Dragon', 'Ice'], 'name': 'White Kyurem', 'move': ['Ice', [140, 90], 's', 'Ice Burn']}, {'hp': 97, 'atk': 167, 'def': 97, 'spatk': 167, 'spdef': 97, 'speed': 129, 'type': ['Psychic', 'Dragon'], 'name': 'Ultra Necrozma', 'move': ['Psychic', [200, 100], 's', 'Light That Burns The Sky']}, {'hp': 97, 'atk': 113, 'def': 109, 'spatk': 157, 'spdef': 127, 'speed': 77, 'type': ['Psychic', 'Ghost'], 'name': 'Dawn Necrozma', 'move': ['Psychic', [200, 100], 's', 'Light That Burns The Sky']}, {'hp': 80, 'atk': 185, 'def': 115, 'spatk': 40, 'spdef': 105, 'speed': 75, 'type': ['Bug', 'Fighting'], 'name': 'Mega Heracross', 'move': ['Bug', [120, 85], 'p', 'Megahorn']}, {'hp': 80, 'atk': 185, 'def': 115, 'spatk': 40, 'spdef': 105, 'speed': 75, 'type': ['Bug', 'Fighting'], 'name': 'Mega Heracross', 'move': ['Bug', [80, 100], 'p', 'X-Scissor']}, {'hp': 106, 'atk': 150, 'def': 70, 'spatk': 194, 'spdef': 120, 'speed': 140, 'type': ['Psychic', ''], 'name': 'Mega Mewtwo Y', 'move': ['Psychic', [120, 100], 's', 'Future sight']}, {'hp': 108, 'atk': 170, 'def': 115, 'spatk': 120, 'spdef': 95, 'speed': 92, 'type': ['Dragon', 'Ground'], 'name': 'Mega Garchomp', 'move': ['Ground', [100, 100], 'p', 'Earthquake']}, {'hp': 105, 'atk': 180, 'def': 100, 'spatk': 180, 'spdef': 100, 'speed': 115, 'type': ['Dragon', 'Flying'], 'name': 'Mega Rayquaza ', 'move': ['Flying', [120, 100], 'p', 'Dragon Ascent']}, {'hp': 128, 'atk': 107, 'def': 125, 'spatk': 73, 'spdef': 105, 'speed': 87, 'type': ['Flying', 'Steel'], 'name': 'gigantamax Corviknight', 'move': ['Flying', [150, 100], 'p', 'G-max Wind Rage']}, {'hp': 100, 'atk': 150, 'def': 120, 'spatk': 120, 'spdef': 100, 'speed': 90, 'type': ['Dragon', 'Electric'], 'name': 'Zekrom', 'move': ['Electric', [100, 100], 'p', 'Fusion Bolt']}, {'hp': 100, 'atk': 150, 'def': 120, 'spatk': 120, 'spdef': 100, 'speed': 90, 'type': ['Dragon', 'Electric'], 'name': 'Zekrom', 'move': ['Electric', [130, 85], 'p', 'Bolt strike']}, {'hp': 95, 'atk': 118, 'def': 90, 'spatk': 134, 'spdef': 90, 'speed': 95, 'type': ['Electric', 'Poison'], 'name': 'gigantamax Toxtricity ', 'move': ['Electric', [150, 100], 's', 'G-max Stun shock']}, {'hp': 70, 'atk': 150, 'def': 140, 'spatk': 65, 'spdef': 100, 'speed': 75, 'type': ['Bug', 'Steel'], 'name': 'Mega Scizor', 'move': ['Bug', [80, 100], 'p', 'X-Scissor']}, {'hp': 350, 'atk': 120, 'def': 75, 'spatk': 270, 'spdef': 160, 'speed': 214, 'type': ['Bug', 'Psychic'], 'name': 'Mega Shadow Mewtwo Y', 'move': ['Shadow', [130, 90], 's', 'Shadow Nova']}, {'hp': 100, 'atk': 125, 'def': 102, 'spatk': 80, 'spdef': 102, 'speed': 95, 'type': ['Poison', ''], 'name': 'gigantamax Garbodor', 'move': ['Poison', [150, 100], 'p', 'G-max Malodor']}, {'hp': 130, 'atk': 100, 'def': 140, 'spatk': 100, 'spdef': 110, 'speed': 50, 'type': ['Rock', 'Fire'], 'name': 'gigantamax Coalossal', 'move': ['Rock', [150, 100], 'p', 'G-max Volcalith']}, {'hp': 115, 'atk': 140, 'def': 130, 'spatk': 55, 'spdef': 55, 'speed': 40, 'type': ['Ground', 'Rock'], 'name': 'Rhyperior', 'move': ['Rock', [150, 90], 'p', 'rock wrecker']}, {'hp': 85, 'atk': 115, 'def': 75, 'spatk': 55, 'spdef': 75, 'speed': 82, 'type': ['Rock', ''], 'name': 'Midnight Lycanroc', 'move': ['Rock', [190, 100], 'p', 'splintered stormshards']}, {'hp': 85, 'atk': 80, 'def': 95, 'spatk': 130, 'spdef': 141, 'speed': 84, 'type': ['Fairy', ''], 'name': 'gigantamax Alcremie', 'move': ['Fairy', [150, 100], 's', 'G-max Finale']}, {'hp': 77, 'atk': 110, 'def': 115, 'spatk': 156, 'spdef': 123, 'speed': 49, 'type': ['Psychic', 'Fairy'], 'name': 'gigantamax Hatterene', 'move': ['Fairy', [150, 100], 's', 'G-max Smite']}]
