--- requirments.txt
@@ -1,15 +1,0 @@
-aiohttp==3.8.4
-aiosignal==1.3.1
-async-timeout==4.0.2
-attrs==23.1.0
-charset-normalizer==3.1.0
-dinteractions-Paginator==2.0.2
-discord-py-interactions==4.4.1
-discord-py-slash-command==4.2.1
-frozenlist==1.3.3
-idna==3.4
-interactions-wait-for==1.0.6
-multidict==6.0.4
-numpy==1.24.3
-opencv-python==4.7.0.72
-yarl==1.9.2
